Shorten items
Thinking it might be easier to read if "Meteor" and "React Native", which are givens for this page, are taken out of items? Also, how is the third Accounts item different from the first? 

Curious why the boilerplate doesn't use the `react-native-meteor` you seem to favor in the FAQ – is it not ready yet?